prepare.libreswan: prevent ipsec restart after each terdown
This should reduce the number of pluto crashes in libreswan5.1 significantly. There is crash happening in pluto_shutdown in random places.

+++ prepare/libreswan.sh
@@ -198,7 +198,7 @@ libreswan_setup ()
         NSS_CLIENT_DIR="/etc/ipsec.d/"
     fi
 
-    if (($LIB_MAJOR >= 5)); then
+    if (($LIB_MAJOR >= 5)) && [ ! -f /etc/ipsec.d/ikev1_accept.conf ]; then
         echo -e "config setup\n  ikev1-policy = accept" > /etc/ipsec.d/ikev1_accept.conf
         # apply the config in the client side
         ipsec restart
